+++ doc/changes.md
@@ -167,6 +167,7 @@ as part of K95 at this time, the default terminal remains VT220 for now.
    `SET TERM CLEAR-ON-RESIZE { ON, OFF }` 
  - Soft-fonts (DRCS) for VT220 and higher emulations are now supported
  - The VT520 Play Sound control sequence is now supported.
+ - Proper Smooth Scroll is now supported in K95G.
 
 ### Enhancements
  - The Control Sequences documentation ([preliminary version available online](https://davidrg.github.io/ckwin/dev/ctlseqs.html))
@@ -253,6 +254,9 @@ as part of K95 at this time, the default terminal remains VT220 for now.
    x86-64 and arm64 hosts to enable better terminal emulation in PTY sessions.
  - The PTY command will now launch the Windows shell (cmd.exe) by default if no
    other command is specified
+ - WYSE WY-50 and compatible escape sequences for setting the smooth scrolling
+   speed (`ESC `` <`, `ESC `` =`, `ESC `` >`, `ESC `` ?`) now set the
+   appropriate smooth-scroll speed.
 
 ### New terminal control sequences
 > [!NOTE]